juntando create_text_file com download_musicas

+++ scripts/scrap_songs.py
@@ -4,7 +4,7 @@ from bs4 import BeautifulSoup
 from langdetect import detect
 
 def get_soup(link):
-    page = requests.get(link)
+    page = requests.get(link)# talvez precise usar 'verify=False' por conta de um 'SSL certificate error'
     soup = BeautifulSoup(page.content, 'html.parser')
     return soup
 
@@ -34,26 +34,27 @@ def scrap_letra_musica(link) -> str:
     song_lyric = soup.find(class_="lyric-original").get_text(separator="\n")
     return song_lyric
 
-def create_text_file(folder_name,file_name,file_content):
-    folder_path = os.path.join(os.getcwd(), folder_name)
-    os.makedirs(folder_path, exist_ok=True)
-    file_path = os.path.join(folder_path, f'{file_name}.txt')
-    with open(file_path, "w") as file:
-        file.write(file_content)
-
 def download_musicas(artist):
     links_dict = scrap_links_musicas(artist)
     downloaded = 0
+    folder_path = f'{os.getcwd()}\\musicas\\{artist}'
+    #cria pasta caso ainda não exista
+    os.makedirs(folder_path, exist_ok=True)
     for song_title in links_dict:
-        folder_name = f'musicas\\{artist}'
         file_name = replace_bad_char(song_title)
+        file_path = f'{folder_path}\\{file_name}.txt'
+        if os.path.exists(file_path):
+            print (f'"{file_name}" já existe')
+            continue
         song_lyric = scrap_letra_musica(links_dict[song_title])
         #só salvar se a música estiver em portugês
-        if detect(song_lyric) == 'pt':
-            create_text_file(folder_name, file_name, song_lyric)
-            downloaded += 1
-        else:
+        if detect(song_lyric) != 'pt':
             print (f'"{song_title}" não foi salva pois não está em português.')
+            continue
+        with open(file_path, "w") as file:
+            print (f'Salvando "{song_title}"')
+            file.write(song_lyric)
+        downloaded += 1
     print (f'{downloaded} músicas foram salvas')
 
 if __name__ == '__main__':
